RaspberryPi Pico support (#152)
<!-- Please Give Your PR a relevant title-->

## Description of Problem
After the loss of multiple Teensy4.1s, we've been looking at some more
affordable alternatives for Catalyst-2. Current candidates include the
ProMicro and RaspberryPi Pico. These boards have slightly different
pinouts and chip architectures than the Teensy, so some changes will
need to be made to the firmware to support them.

## Solution
- Added a PlatformIO `env` for the RaspberryPi Pico
- After migrating some chip settings to a central settings file, a build
flag was added to the `pico` env to separate settings for different
architectures.
- Settings for the RaspberryPi Pico were added to support a different
pinout and reduced RAM.

**A Raspberry Pi Pico should now be a drop-in replacement for the Teensy
in Catalyst-2**